Major Model Refactoring
I needed to refactor the whole model as I found that it was not generalizing well at all to the test set. Therefore, I decided to merge the DATA and TEST files at the start into one big dataset, then partitioned the dataset in later steps into training, validation and accuracy.